small bugs, removed old clientHandler handling programmes/ modules /classes differently now
